HTTP Payload Validation
Add validation logic for Job and JobInfo

A new validation framework is introduced to validate job and job info entities. To achieve this, new Validator[A] typeclass and its instances for Job and JobInfo are added. This is integrated into request handling in JobRoutes where incoming requests are validated before processing. The validation-related logic is abstracted into their own files for separation of concerns.